+++ src/docs/components/switch/api.ts
@@ -1,6 +1,6 @@
-import { use, self, type APIReference } from '$site/index.js';
+import { use, self, type ComponentReference } from '$site/index.js';
 
-const switchAPI: APIReference = {
+const switchAPI: ComponentReference = {
 	name: 'Switch',
 	props: [
 		{