+++ src/lib/components/RadioGroup/RadioGroupItem.svelte
@@ -28,5 +28,5 @@
 </script>
 
 <button bind:this={self} use:useActions={use} class={classProp(klass, ctx.state)} {...ctx.attrs} {...props}>
-	{@render children(ctx.state)}
+	{@render children?.(ctx.state)}
 </button>